Add the compression type for the oap data file name if it's compresed. (#314)
* Add the compression type for the oap data file name

If the oap is uncompressed, keep the original file name.
Verified that the data file name is changed correctly as the
compression type is changed dynamically.

* Nothing to do if the oap data file is uncompressed.

If the oap data file is uncompressed, keep the original file name.

* Avoid to use var

No need var in order to add the compression type to the data file name

* Be more concise with the functional programming style

* trim the whitespaces if any

* Add one test case to test if the compression type is added as expected

* Keep concise with scala style

* Use foreach to test each compression types

* Following the scala style to be more concise