Remove OapIndexInfo HeartBeat (#1069)
We collect OapIndexInfo and report it to driver as a HeartBeat, then only print a log at driver side and no other use. And this heartbeat is too many times, it will bring some pressure to the event handling of Driver, so delete it now. We need to consider other ways of collect index usage information.