SQL grammar decouple (#1115)
Add a new OapSqlBase.g4 file which only contains the grammar of OAP. Redefine the OapSparkSqlParser so that we can parse the SQL by using custom SqlParser first. If it fails, then use SparkSqlParser to parse SQL. Remove SqlBase.g4 which rely on the source code of Spark SQL. We inject custom SqlParser on OapExtensions class.